copd-app/obs_generator/requirements.txt
fastapi[standard]
fhir.resources
pytz
convert
